Separated logix & UX stuff
